requirements.txt
requests==2.31.0
openai==2.8.1
python-dotenv==1.0.0
Pillow==10.2.0
